feat : Use scikit-learn and Keras wrapper for cross validation and evaluation on AMNWt model
Use the build_model method into a scikit-learn cross validation code. Define the two scorers correrponding to the AMN loss and the R2 score. This code is aim to remplace train_evaluate method. It does exacly the same, using the Keras, Scikit-learn API, expliciting the specificity of the project. The only change here is that, as recomended we do not overfiting anymore on the training set by taking the best model on cross-validation.